JavaClient restructure complete with two helper classes
